+++ lib/src/commonMain/kotlin/ru/kode/pomskykt/Expr.kt
@@ -36,7 +36,7 @@ import ru.kode.pomskykt.visitor.walkRule
 class Expr(val rule: Rule) {
 
     companion object {
-        const val VERSION = "0.17.0"
+        const val VERSION = "0.18.0"
 
         /**
          * Parse a pomsky expression.